+++ src/table/boot.rs
@@ -1,6 +1,6 @@
 //! UEFI services available during boot.
 
-use super::Header;
+use super::{Header, Revision};
 use crate::data_types::Align;
 use crate::proto::device_path::{DevicePath, FfiDevicePath};
 #[cfg(feature = "exts")]
@@ -14,6 +14,7 @@ use core::cell::UnsafeCell;
 use core::ffi::c_void;
 use core::fmt::{Debug, Formatter};
 use core::mem::{self, MaybeUninit};
+use core::ops::{Deref, DerefMut};
 use core::ptr::NonNull;
 use core::{ptr, slice};
 
@@ -437,6 +438,9 @@ impl BootServices {
     /// More than one event of type `EventType::TIMER` may be part of a single event group. However,
     /// there is no mechanism for determining which of the timers was signaled.
     ///
+    /// This operation is only supported starting with UEFI 2.0; earlier
+    /// versions will fail with [`Status::UNSUPPORTED`].
+    ///
     /// # Safety
     ///
     /// The caller must ensure they are passing a valid `Guid` as `event_group`, if applicable.
@@ -448,6 +452,10 @@ impl BootServices {
         notify_ctx: Option<NonNull<c_void>>,
         event_group: Option<NonNull<Guid>>,
     ) -> Result<Event> {
+        if self.header.revision < Revision::EFI_2_00 {
+            return Err(Status::UNSUPPORTED.into());
+        }
+
         let mut event = MaybeUninit::<Event>::uninit();
 
         (self.create_event_ex)(
@@ -568,14 +576,21 @@ impl BootServices {
     /// protections must be implemented by user-level code, for example via a
     /// global `HashSet`.
     ///
+    /// # Safety
+    ///
+    /// This method is unsafe because the handle database is not
+    /// notified that the handle and protocol are in use; there is no
+    /// guarantee that they will remain valid for the duration of their
+    /// use. Use [`open_protocol`] instead.
+    ///
     /// [`open_protocol`]: BootServices::open_protocol
     #[deprecated(note = "it is recommended to use `open_protocol` instead")]
-    pub fn handle_protocol<P: ProtocolPointer + ?Sized>(
+    pub unsafe fn handle_protocol<P: ProtocolPointer + ?Sized>(
         &self,
         handle: Handle,
     ) -> Result<&UnsafeCell<P>> {
         let mut ptr = ptr::null_mut();
-        (self.handle_protocol)(handle, &P::GUID, &mut ptr).into_with_val(|| unsafe {
+        (self.handle_protocol)(handle, &P::GUID, &mut ptr).into_with_val(|| {
             let ptr = P::mut_ptr_from_ffi(ptr) as *const UnsafeCell<P>;
             &*ptr
         })
@@ -639,6 +654,57 @@ impl BootServices {
                 },
             )
         }
+    }
+
+    /// Find an arbitrary handle that supports a particular
+    /// [`Protocol`]. Returns [`NOT_FOUND`] if no handles support the
+    /// protocol.
+    ///
+    /// This method is a convenient wrapper around
+    /// [`BootServices::locate_handle_buffer`] for getting just one
+    /// handle. This is useful when you don't care which handle the
+    /// protocol is opened on. For example, [`DevicePathToText`] isn't
+    /// tied to a particular device, so only a single handle is expected
+    /// to exist.
+    ///
+    /// [`NOT_FOUND`]: Status::NOT_FOUND
+    /// [`DevicePathToText`]: uefi::proto::device_path::text::DevicePathToText
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use uefi::proto::device_path::text::DevicePathToText;
+    /// use uefi::table::boot::{BootServices, OpenProtocolAttributes, OpenProtocolParams};
+    /// use uefi::Handle;
+    /// # use uefi::Result;
+    ///
+    /// # fn get_fake_val<T>() -> T { todo!() }
+    /// # fn test() -> Result {
+    /// # let boot_services: &BootServices = get_fake_val();
+    /// # let image_handle: Handle = get_fake_val();
+    /// let handle = boot_services.get_handle_for_protocol::<DevicePathToText>()?;
+    /// let device_path_to_text = boot_services.open_protocol::<DevicePathToText>(
+    ///     OpenProtocolParams {
+    ///         handle,
+    ///         agent: image_handle,
+    ///         controller: None,
+    ///     },
+    ///     OpenProtocolAttributes::Exclusive,
+    /// )?;
+    /// # Ok(())
+    /// # }
+    /// ```
+    pub fn get_handle_for_protocol<P: Protocol>(&self) -> Result<Handle> {
+        // Delegate to a non-generic function to potentially reduce code size.
+        self.get_handle_for_protocol_impl(&P::GUID)
+    }
+
+    fn get_handle_for_protocol_impl(&self, guid: &Guid) -> Result<Handle> {
+        self.locate_handle_buffer(SearchType::ByProtocol(guid))?
+            .handles()
+            .first()
+            .cloned()
+            .ok_or_else(|| Status::NOT_FOUND.into())
     }
 
     /// Load an EFI image into memory and return a [`Handle`] to the image.
@@ -877,6 +943,8 @@ impl BootServices {
         )
         .into_with_val(|| unsafe {
             let interface = P::mut_ptr_from_ffi(interface) as *const UnsafeCell<P>;
+
+            #[allow(deprecated)]
             ScopedProtocol {
                 interface: &*interface,
                 open_params: params,
@@ -951,9 +1019,18 @@ impl BootServices {
     /// Returns a protocol implementation, if present on the system.
     ///
     /// The caveats of `BootServices::handle_protocol()` also apply here.
-    pub fn locate_protocol<P: ProtocolPointer + ?Sized>(&self) -> Result<&UnsafeCell<P>> {
+    ///
+    /// # Safety
+    ///
+    /// This method is unsafe because the handle database is not
+    /// notified that the handle and protocol are in use; there is no
+    /// guarantee that they will remain valid for the duration of their
+    /// use. Use [`BootServices::get_handle_for_protocol`] and
+    /// [`BootServices::open_protocol`] instead.
+    #[deprecated(note = "it is recommended to use `open_protocol` instead")]
+    pub unsafe fn locate_protocol<P: ProtocolPointer + ?Sized>(&self) -> Result<&UnsafeCell<P>> {
         let mut ptr = ptr::null_mut();
-        (self.locate_protocol)(&P::GUID, ptr::null_mut(), &mut ptr).into_with_val(|| unsafe {
+        (self.locate_protocol)(&P::GUID, ptr::null_mut(), &mut ptr).into_with_val(|| {
             let ptr = P::mut_ptr_from_ffi(ptr) as *const UnsafeCell<P>;
             &*ptr
         })
@@ -1023,21 +1100,17 @@ impl BootServices {
             },
             OpenProtocolAttributes::Exclusive,
         )?;
-        let loaded_image = unsafe { &*loaded_image.interface.get() };
-
-        let device_handle = loaded_image.device();
 
         let device_path = self.open_protocol::<DevicePath>(
             OpenProtocolParams {
-                handle: device_handle,
+                handle: loaded_image.device(),
                 agent: image_handle,
                 controller: None,
             },
             OpenProtocolAttributes::Exclusive,
         )?;
-        let mut device_path = unsafe { &*device_path.interface.get() };
 
-        let device_handle = self.locate_device_path::<SimpleFileSystem>(&mut device_path)?;
+        let device_handle = self.locate_device_path::<SimpleFileSystem>(&mut &*device_path)?;
 
         self.open_protocol::<SimpleFileSystem>(
             OpenProtocolParams {
@@ -1323,6 +1396,7 @@ pub struct OpenProtocolParams {
 /// protocol and why [`UnsafeCell`] is used.
 pub struct ScopedProtocol<'a, P: Protocol + ?Sized> {
     /// The protocol interface.
+    #[deprecated(since = "0.16.0", note = "use Deref and DerefMut instead")]
     pub interface: &'a UnsafeCell<P>,
 
     open_params: OpenProtocolParams,
@@ -1343,6 +1417,26 @@ impl<'a, P: Protocol + ?Sized> Drop for ScopedProtocol<'a, P> {
         // and the error can't be propagated out of drop anyway, so just
         // assert success.
         assert_eq!(status, Status::SUCCESS);
+    }
+}
+
+impl<'a, P: Protocol + ?Sized> Deref for ScopedProtocol<'a, P> {
+    type Target = P;
+
+    fn deref(&self) -> &Self::Target {
+        #[allow(deprecated)]
+        unsafe {
+            &*self.interface.get()
+        }
+    }
+}
+
+impl<'a, P: Protocol + ?Sized> DerefMut for ScopedProtocol<'a, P> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        #[allow(deprecated)]
+        unsafe {
+            &mut *self.interface.get()
+        }
     }
 }